+++ projects/dxt-lossless-transform-bc2/benches/transform_standard/avx512.rs
@@ -1,5 +1,7 @@
 use criterion::{black_box, BenchmarkId};
-use dxt_lossless_transform_bc2::transforms::standard::transform::{permute_512, permute_512_v2};
+use dxt_lossless_transform_bc2::transforms::standard::transform::bench::{
+    permute_512, permute_512_v2,
+};
 use safe_allocator_api::RawAlloc;
 
 fn bench_permute(b: &mut criterion::Bencher, input: &RawAlloc, output: &mut RawAlloc) {
